+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/mainnet/TimestampScheduleBuilder.java
@@ -33,15 +33,8 @@ public class TimestampScheduleBuilder extends AbstractProtocolScheduleBuilder {
       final ProtocolSpecAdapters protocolSpecAdapters,
       final PrivacyParameters privacyParameters,
       final boolean isRevertReasonEnabled,
-      final boolean quorumCompatibilityMode,
       final EvmConfiguration evmConfiguration) {
-    super(
-        config,
-        protocolSpecAdapters,
-        privacyParameters,
-        isRevertReasonEnabled,
-        quorumCompatibilityMode,
-        evmConfiguration);
+    super(config, protocolSpecAdapters, privacyParameters, isRevertReasonEnabled, evmConfiguration);
     this.defaultChainId = Optional.of(defaultChainId);
   }