+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/storage/keyvalue/KeyValueStorageProviderBuilder.java
@@ -30,7 +30,6 @@ public class KeyValueStorageProviderBuilder {
   private KeyValueStorageFactory storageFactory;
   private BesuConfiguration commonConfiguration;
   private MetricsSystem metricsSystem;
-  private boolean isGoQuorumCompatibilityMode;
 
   public KeyValueStorageProviderBuilder withStorageFactory(
       final KeyValueStorageFactory storageFactory) {
@@ -49,12 +48,6 @@ public class KeyValueStorageProviderBuilder {
     return this;
   }
 
-  public KeyValueStorageProviderBuilder isGoQuorumCompatibilityMode(
-      final boolean isGoQuorumCompatibilityMode) {
-    this.isGoQuorumCompatibilityMode = isGoQuorumCompatibilityMode;
-    return this;
-  }
-
   public KeyValueStorageProvider build() {
     checkNotNull(storageFactory, "Cannot build a storage provider without a storage factory.");
     checkNotNull(
@@ -65,26 +58,13 @@ public class KeyValueStorageProviderBuilder {
     final KeyValueStorage worldStatePreImageStorage =
         new LimitedInMemoryKeyValueStorage(DEFAULT_WORLD_STATE_PRE_IMAGE_CACHE_SIZE);
 
-    final KeyValueStorage privateWorldStatePreImageStorage =
-        new LimitedInMemoryKeyValueStorage(DEFAULT_WORLD_STATE_PRE_IMAGE_CACHE_SIZE);
-
     // this tickles init needed for isSegmentIsolationSupported
     storageFactory.create(KeyValueSegmentIdentifier.BLOCKCHAIN, commonConfiguration, metricsSystem);
-    if (isGoQuorumCompatibilityMode) {
-      return new GoQuorumKeyValueStorageProvider(
-          segment -> storageFactory.create(segment, commonConfiguration, metricsSystem),
-          worldStatePreImageStorage,
-          privateWorldStatePreImageStorage,
-          storageFactory.isSegmentIsolationSupported(),
-          (ObservableMetricsSystem) metricsSystem);
-    } else {
-      return new KeyValueStorageProvider(
-          segment -> storageFactory.create(segment, commonConfiguration, metricsSystem),
-          worldStatePreImageStorage,
-          privateWorldStatePreImageStorage,
-          storageFactory.isSegmentIsolationSupported(),
-          storageFactory.isSnapshotIsolationSupported(),
-          (ObservableMetricsSystem) metricsSystem);
-    }
+    return new KeyValueStorageProvider(
+        segment -> storageFactory.create(segment, commonConfiguration, metricsSystem),
+        worldStatePreImageStorage,
+        storageFactory.isSegmentIsolationSupported(),
+        storageFactory.isSnapshotIsolationSupported(),
+        (ObservableMetricsSystem) metricsSystem);
   }
 }